refactor(ui): separate lib and utils
