+++ platform/packages/ui/src/components/providers/UIProvider.tsx
@@ -1,0 +1,18 @@
+'use client'
+
+import { ThemeProvider as NextThemesProvider } from 'next-themes'
+import * as React from 'react'
+
+export function UIProvider({ children }: { children: React.ReactNode }) {
+  return (
+    <NextThemesProvider
+      attribute="class"
+      defaultTheme="system"
+      enableSystem
+      disableTransitionOnChange
+      enableColorScheme
+    >
+      {children}
+    </NextThemesProvider>
+  )
+}
