+++ platform/packages/ui/src/lib/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { cn } from '@botfi/ui/lib/utils'
+import { cn } from '@botfi/ui/utils'
 
 function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
   return (
