+++ platform/packages/ui/src/lib/separator.tsx
@@ -3,7 +3,7 @@
 import * as SeparatorPrimitive from '@radix-ui/react-separator'
 import * as React from 'react'
 
-import { cn } from '@botfi/ui/lib/utils'
+import { cn } from '@botfi/ui/utils'
 
 function Separator({
   className,
