+++ platform/packages/ui/src/lib/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@botfi/ui/lib/utils'
+import { cn } from '@botfi/ui/utils'
 
 function Skeleton({ className, ...props }: React.ComponentProps<'div'>) {
   return <div data-slot="skeleton" className={cn('bg-accent animate-pulse rounded-md', className)} {...props} />
